Test rasa init.
