+++ tests/cli/test_init.py
@@ -1,25 +1,21 @@
-import pytest
-
-# we reuse a bit of pytest's own testing machinery, this should eventually come
-# from a separatedly installable pytest-cli plugin.
-pytest_plugins = ["pytester"]
+import os
 
 
-@pytest.fixture
-def run(testdir):
-    def do_run(*args):
-        args = ["rasa"] + list(args)
-        return testdir._run(*args)
+def test_init(run):
+    result, temp_dir = run("init", "--no-prompt", "--quiet")
 
-    return do_run
-
-
-def test_init(tmpdir, run):
-    run("init", "--no-prompt")
+    assert os.path.exists(os.path.join(temp_dir, "actions.py"))
+    assert os.path.exists(os.path.join(temp_dir, "domain.yml"))
+    assert os.path.exists(os.path.join(temp_dir, "config.yml"))
+    assert os.path.exists(os.path.join(temp_dir, "credentials.yml"))
+    assert os.path.exists(os.path.join(temp_dir, "endpoints.yml"))
+    assert os.path.exists(os.path.join(temp_dir, "models"))
+    assert os.path.exists(os.path.join(temp_dir, "data", "nlu.md"))
+    assert os.path.exists(os.path.join(temp_dir, "data", "stories.md"))
 
 
 def test_init_help(run):
-    help = run("init", "--help")
+    help, _ = run("init", "--help")
 
     assert (
         help.outlines[0] == "usage: rasa init [-h] [-v] [-vv] [--quiet] [--no-prompt]"
